+++ src/events/OnFavoriteEvent.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <Geode/loader/Event.hpp>
+
+using namespace geode::prelude;
+
+class OnFavoriteEvent : public Event {
+public:
+    OnFavoriteEvent() = default;
+};